Move test_string() into test_code_format to check compiled code.
